+++ packages/@headlessui-react/src/internal/focus-sentinel.tsx
@@ -1,6 +1,6 @@
 import React, { useState, FocusEvent as ReactFocusEvent } from 'react'
 
-import { VisuallyHidden } from './visually-hidden'
+import { Hidden } from './hidden'
 
 interface FocusSentinelProps {
   onFocus(): boolean
@@ -12,7 +12,7 @@ export function FocusSentinel({ onFocus }: FocusSentinelProps) {
   if (!enabled) return null
 
   return (
-    <VisuallyHidden
+    <Hidden
       as="button"
       type="button"
       onFocus={(event: ReactFocusEvent) => {
